Move modalActivityBlock component to activity-blocks folder
